% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcr_check_dataset_is_complete.R
\name{spcr_check_dataset_is_complete}
\alias{spcr_check_dataset_is_complete}
\title{Check all required data items are provided}
\usage{
spcr_check_dataset_is_complete(report_config, measure_data)
}
\arguments{
\item{report_config}{A data frame. The report config detailing required report items}

\item{measure_data}{Data frame in either wide format}
}
\value{
logical TRUE if check is successful, else an error message
}
\description{
Check all required data items are provided
}
