% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPCreporter-package.R
\docType{package}
\name{SPCreporter-package}
\alias{SPCreporter}
\alias{SPCreporter-package}
\title{SPCreporter: Creates Metric Reports using Statistical Process Control in the NHS style}
\description{
Takes a dataset file and a configuration file to produce an html report with each metric presented as an SPC chart. The initial rendering of the report contains the charts folded to enable quick scanning of "variation" and "assurance" status. Each metric can be unfolded to reveal the SPC chart to support clinical and management planning. The NHSRplothtedots package is used for SPC calcuations, and the "variation" and "assurance" is in line with the popular standard within the UK's National Health Service (NHS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThomUK/SPCreporter}
  \item Report bugs at \url{https://github.com/ThomUK/SPCreporter/issues}
}

}
\author{
\strong{Maintainer}: Tom Smith \email{tomsmith_uk@hotmail.com}

}
\keyword{internal}
