% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcr_make_report.R
\name{with_ext}
\alias{with_ext}
\title{Copied from {xfun}
https://github.com/yihui/xfun/blob/main/R/paths.R}
\usage{
with_ext(x, ext, extra = "")
}
\arguments{
\item{x}{A character of file paths.}

\item{ext}{A vector of new extensions. It must be either of length 1, or the
same length as \code{x}.}

\item{extra}{Extra characters to be allowed in the extensions. By default,
only alphanumeric characters are allowed (and also some special cases in
\sQuote{Details}). If other characters should be allowed, they can be
specified in a character string, e.g., \code{"-+!_#"}.}
}
\description{
Copied from {xfun}
https://github.com/yihui/xfun/blob/main/R/paths.R
}
