% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcr_check_dataset_is_complete.R
\name{spcr_check_dataset_is_complete}
\alias{spcr_check_dataset_is_complete}
\title{Check all required data items are provided}
\usage{
spcr_check_dataset_is_complete(report_config, measure_data)
}
\arguments{
\item{report_config}{dataframe. The report config detailing required report items}

\item{measure_data}{list. The data items to check}
}
\value{
logical
}
\description{
Check all required data items are provided
}
