% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcr_make_data_bundle.R
\name{spcr_make_data_bundle}
\alias{spcr_make_data_bundle}
\title{Make a bundle of data including all SPC calcs}
\usage{
spcr_make_data_bundle(measure_data, report_config, measure_config)
}
\arguments{
\item{measure_data}{list. List containing dataframes of data in wide format}

\item{report_config}{dataframe. Config information for the report}

\item{measure_config}{dataframe. Config information for the measures}
}
\value{
dataframe. A nested dataframe containing calculated charts and parsed text
}
\description{
Make a bundle of data including all SPC calcs
}
