% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcr_make_data_bundle.R
\name{spcr_make_data_bundle}
\alias{spcr_make_data_bundle}
\title{Make a bundle of data}
\usage{
spcr_make_data_bundle(
  measure_data = test_measure_data,
  report_config = test_report_config,
  measure_config = test_measure_config
)
}
\arguments{
\item{measure_data}{list. List containing data frames of data in wide format}

\item{report_config}{data frame. Config information for the report}

\item{measure_config}{data frame. Config information for the measures}
}
\value{
data frame. A nested data frame containing source data for the report
}
\description{
Make a bundle of data
}
