% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcr_make_report.R
\name{spcr_make_report}
\alias{spcr_make_report}
\title{Make Report}
\usage{
spcr_make_report(
  data_bundle,
  report_title = "SPC Report",
  output_directory = "/"
)
}
\arguments{
\item{data_bundle}{list. The pre-processed bundle of information (made with spcr_make_data_bundle())}

\item{report_title}{String. The title of the report}

\item{output_directory}{String. The name of the directory to save the resulting report to}
}
\description{
Make Report
}
